consolidate interfaces into GeoProviders.ts, improve deprecation messages
