+++ core/frontend/src/BingLocation.ts
@@ -10,11 +10,11 @@ import { Cartographic } from "@itwin/core-common";
 import { request } from "./request/Request";
 import { IModelApp } from "./IModelApp";
 import { GlobalLocation } from "./ViewGlobalLocation";
-import type { LocationProvider } from "./LocationProvider";
+import type { LocationProvider } from "./GeoProviders";
 
 /** Provides an interface to the [Bing Maps location services](https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/).
  * Use of this service requires an API key to be supplied via [[MapLayerOptions.BingMaps]] in the [[IModelAppOptions.mapLayerOptions]] passed to [[IModelApp.startup]].
- * @deprecated in 5.11.0. Use [[LocationProvider]] via [[IModelAppOptions.locationProvider]].
+ * @deprecated in 5.11.0. Provide a [[LocationProvider]] implementation via [[IModelAppOptions.locationProvider]].
  * @public
  * @extensions
  */
