+++ core/frontend/src/IModelApp.ts
@@ -41,9 +41,7 @@ import * as spatialViewState from "./SpatialViewState";
 import { TentativePoint } from "./TentativePoint";
 import { RealityDataSourceProviderRegistry } from "./RealityDataSource";
 import { BingElevationProvider, MapLayerFormatRegistry, MapLayerOptions, TerrainProviderRegistry, TileAdmin } from "./tile/internal";
-import { ElevationProvider } from "./ElevationProvider";
-import { GeoidProvider } from "./GeoidProvider";
-import { LocationProvider } from "./LocationProvider";
+import { ElevationProvider, GeoidProvider, LocationProvider } from "./GeoProviders";
 import { BingLocationProvider } from "./BingLocation";
 import * as accudrawTool from "./tools/AccuDrawTool";
 import * as clipViewTool from "./tools/ClipViewTool";
