--- core/frontend/src/LocationProvider.ts
@@ -1,17 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-/** @packageDocumentation
- * @module Tiles
- */
-
-import { GlobalLocation } from "./ViewGlobalLocation";
-
-/** Provides geocoding — converting a query string to a geographic location.
- * @beta
- */
-export interface LocationProvider {
-  /** Return the location for a query string, or undefined if not found. */
-  getLocation(query: string): Promise<GlobalLocation | undefined>;
-}
