+++ core/frontend/src/core-frontend.ts
@@ -8,9 +8,7 @@ export * from "./AccuSnap";
 export * from "./AuxCoordSys";
 export * from "./BingLocation";
 export * from "./BriefcaseConnection";
-export * from "./ElevationProvider";
-export * from "./GeoidProvider";
-export * from "./LocationProvider";
+export * from "./GeoProviders";
 export * from "./BriefcaseTxns";
 export * from "./CatalogConnection";
 export * from "./CategorySelectorState";
